Trekk ut hjepletekst44px. Gjem hjelpeteksten på panelet
